Added the ability to change the tab bar position to inspector sidebar.
