+++ CodeEdit/Features/InspectorSidebar/InspectorSidebarView.swift
@@ -22,6 +22,8 @@ struct InspectorSidebarView: View {
         self.workspace = workspace
     }
 
+    @AppSettings(\.general.inspectorTabBarPosition) var sidebarPosition: SettingsData.SidebarTabBarPosition
+
     var body: some View {
         VStack {
             if let path = tabManager.activeTabGroup.selected?.fileDocument?.fileURL?.path(percentEncoded: false) {
@@ -45,6 +47,7 @@ struct InspectorSidebarView: View {
                 NoSelectionInspectorView()
             }
         }
+        .clipShape(Rectangle())
         .frame(
             minWidth: CodeEditWindowController.minSidebarWidth,
             idealWidth: 260,
@@ -52,9 +55,17 @@ struct InspectorSidebarView: View {
             maxHeight: .infinity,
             alignment: .top
         )
+        .safeAreaInset(edge: .trailing, spacing: 0) {
+            if sidebarPosition == .side {
+                InspectorSidebarTabBar(selection: $selection, position: sidebarPosition)
+            }
+        }
         .safeAreaInset(edge: .top, spacing: 0) {
-            InspectorSidebarToolbarTop(selection: $selection)
-                .background(EffectView(.windowBackground))
+            if sidebarPosition == .top {
+                InspectorSidebarTabBar(selection: $selection, position: sidebarPosition)
+            } else {
+                Divider()
+            }
         }
     }
 }
